refactor after dev 2
