+++ src/services/actions/login-user.js
@@ -1,0 +1,25 @@
+import { createAsyncThunk } from '@reduxjs/toolkit'
+import { AuthService } from '../../utils/auth-service'
+import { getAccessToken } from '../reducers/auth-user-slice'
+
+export const loginUser = createAsyncThunk(
+	'authUserSlice/loginUser',
+	async function (data, { rejectWithValue }) {
+		const { userEmail, userPassword } = data
+		try {
+			const res = await AuthService.login(userEmail, userPassword)
+			const {
+				accessToken,
+				refreshToken,
+				user: { name }
+			} = res.data
+			localStorage.setItem('refreshToken', refreshToken)
+			localStorage.setItem('accessToken', getAccessToken(accessToken))
+			return { email: userEmail, name: name }
+		} catch (e) {
+			localStorage.setItem('refreshToken', '')
+			localStorage.setItem('accessToken', '')
+			return rejectWithValue('Некорректная почта или пароль')
+		}
+	}
+)
